// src/app/model/user.ts
export class User {
    id?: string = '';
    username: string = '';
    email?: string = '';
    password?: string = '';
    firstName?: string = '';
    lastName?: string = '';
    slug?: string = '';
    tags?: string[] = [];
    dateOfBirth?: string = '';
    city?: string = '';
    country?: string = '';
}